Integrated Topshelf into the solution turning IrcExploitChecker.App into a Windows service application.
