+++ Projects/ExploitChecker/AbusiveNickExploitChecker.cs
@@ -17,7 +17,7 @@ namespace ExploitChecker
             this.nicksFunc = nicksFunc;
         }
 
-        protected Func<IEnumerable<string>> NicksFunc
+        protected virtual Func<IEnumerable<string>> NicksFunc
         {
             get
             {
