+++ Projects/ExploitChecker/BanListExploitChecker.cs
@@ -40,7 +40,7 @@ namespace ExploitChecker
             }
         }
 
-        protected List<string> BanList
+        protected virtual List<string> BanList
         {
             get
             {
@@ -50,12 +50,12 @@ namespace ExploitChecker
 
         public abstract IEnumerable<ExploitType> Check(Func<string> tuple);
 
-        public bool Contains(string entry)
+        public virtual bool Contains(string entry)
         {
             return BanList.Contains(entry);
         }
 
-        public void Add(string entry)
+        public virtual void Add(string entry)
         {
             if (!Contains(entry))
             {
@@ -63,7 +63,7 @@ namespace ExploitChecker
             }
         }
 
-        public void Remove(string entry)
+        public virtual void Remove(string entry)
         {
             if (Contains(entry))
             {
@@ -71,13 +71,12 @@ namespace ExploitChecker
             }
         }
 
-
-        public IEnumerable<string> GetAll()
+        public virtual IEnumerable<string> GetAll()
         {
             return BanList.ToArray();
         }
 
-        public void SetAll(IEnumerable<string> banList)
+        public virtual void SetAll(IEnumerable<string> banList)
         {
             BanList.Clear();
             BanList.AddRange(banList);
