+++ Projects/ExploitChecker/ExploitChecker.csproj
@@ -35,12 +35,6 @@
     <StartupObject />
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="ChatSharp">
-      <HintPath>..\packages\ChatSharp.1.0.0\lib\net45\ChatSharp.dll</HintPath>
-    </Reference>
-    <Reference Include="CommandLine">
-      <HintPath>..\packages\CommandLineParser.1.9.71\lib\net45\CommandLine.dll</HintPath>
-    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
@@ -63,7 +57,7 @@
     <Compile Include="AbusiveNickExploitChecker.cs" />
     <Compile Include="LongNickExploitChecker.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TorSectoorChecker.cs" />
+    <Compile Include="TorSectoorExploitChecker.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
