+++ Projects/ExploitChecker/IPAddressExploitChecker.cs
@@ -23,7 +23,7 @@ namespace ExploitChecker
             return exploitTypes;
         }
 
-        private IEnumerable<IPAddress> ResolveIPAddresses(string address)
+        protected virtual IEnumerable<IPAddress> ResolveIPAddresses(string address)
         {
             var ipAddresses = new List<IPAddress>();
 
@@ -50,7 +50,7 @@ namespace ExploitChecker
             return ipAddresses;
         }
 
-        private IEnumerable<ExploitType> TestIPAddresses(IEnumerable<IPAddress> ipAddresses)
+        protected virtual IEnumerable<ExploitType> TestIPAddresses(IEnumerable<IPAddress> ipAddresses)
         {
             var exploitTypes = new List<ExploitType>();
 
@@ -64,6 +64,28 @@ namespace ExploitChecker
             return exploitTypes;
         }
 
-        protected abstract IEnumerable<ExploitType> TestIPAddress(byte[] addressBytes);
+        protected virtual IEnumerable<ExploitType> TestIPAddress(byte[] addressBytes)
+        {
+            var address = FormatAddress(addressBytes);
+
+            try
+            {
+                var hostAddresses = Dns.GetHostAddresses(address);
+
+                var exploitTypes = CheckHostAddresses(hostAddresses);
+
+                return exploitTypes.ToArray();
+            }
+            catch
+            {
+                // TODO : log exception -- Diabolic 15/03/2015
+
+                return Enumerable.Empty<ExploitType>();
+            }
+        }
+
+        protected abstract string FormatAddress(byte[] addressBytes);
+
+        protected abstract IEnumerable<ExploitType> CheckHostAddresses(IPAddress[] hostAddresses);
     }
 }
