+++ Projects/ExploitChecker/LongNickExploitChecker.cs
@@ -16,7 +16,7 @@ namespace ExploitChecker
             this.nickLength = nickLength;
         }
 
-        protected int NickLength
+        protected virtual int NickLength
         {
             get
             {
@@ -24,7 +24,7 @@ namespace ExploitChecker
             }
         }
 
-        public IEnumerable<ExploitType> Check(Func<string> tuple)
+        public virtual IEnumerable<ExploitType> Check(Func<string> tuple)
         {
             if (tuple == null)
             {
