+++ Projects/ExploitChecker/TorSectoorExploitChecker.cs
@@ -15,9 +15,9 @@ namespace ExploitChecker
         {
         }
 
-        protected override IEnumerable<ExploitType> TestIPAddress(byte[] addressBytes)
+        protected override string FormatAddress(byte[] addressBytes)
         {
-            var address = string.Format
+            return string.Format
             (
                 "{0}.{1}.{2}.{3}.tor.dnsbl.sectoor.de",
                 Convert.ToInt32(addressBytes[3]),
@@ -25,24 +25,14 @@ namespace ExploitChecker
                 Convert.ToInt32(addressBytes[1]),
                 Convert.ToInt32(addressBytes[0])
             );
+        }
 
-            try
-            {
-                var hostAddresses = Dns.GetHostAddresses(address);
-
-                var exploitTypes = hostAddresses
+        protected override IEnumerable<ExploitType> CheckHostAddresses(IPAddress[] hostAddresses)
+        {
+            return hostAddresses
                     .Select(x => Convert.ToInt32(x.GetAddressBytes()[3]))
                     .Where(x => x == 2)
                     .Select(x => ExploitType.TorSectoor_Blacklisted);
-
-                return exploitTypes.ToArray();
-            }
-            catch
-            {
-                // TODO : log exception -- Diabolic 15/03/2015
-
-                return Enumerable.Empty<ExploitType>();
-            }
         }
     }
 }
