+++ Projects/IrcExploitChecker.App/App.config
@@ -27,7 +27,7 @@
                 <value>5</value>
             </setting>
             <setting name="Network" serializeAs="String">
-                <value>irc.dal.net</value>
+                <value>eu.undernet.org</value>
             </setting>
             <setting name="WatchedChannels" serializeAs="String">
                 <value>#intervention</value>
