+++ Projects/IrcExploitChecker.App/Properties/Settings.settings
@@ -18,7 +18,7 @@
       <Value Profile="(Default)">5</Value>
     </Setting>
     <Setting Name="Network" Type="System.String" Scope="User">
-      <Value Profile="(Default)">irc.dal.net</Value>
+      <Value Profile="(Default)">eu.undernet.org</Value>
     </Setting>
     <Setting Name="WatchedChannels" Type="System.String" Scope="User">
       <Value Profile="(Default)">#intervention</Value>
