+++ Projects/IrcExploitChecker.App/packages.config
@@ -6,4 +6,5 @@
   <package id="Rx-Linq" version="2.2.5" targetFramework="net45" />
   <package id="Rx-Main" version="2.2.5" targetFramework="net45" />
   <package id="Rx-PlatformServices" version="2.2.5" targetFramework="net45" />
+  <package id="Topshelf" version="3.1.4" targetFramework="net45" />
 </packages>
